requirement.txt
Django==2.2.5
django-pure-pagination==0.3.0
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==2.6.3
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
requests==2.20.1
# mysql-connector
# mysqlclient
